smartjob_full/parser-service/requirements.txt
flask
spacy
pdfminer.six
python-docx
flask-cors
